improve math logic in cleanup

+++ cron/cleanmessagelog.php
@@ -20,13 +20,21 @@ foreach ($settings->getMessagelogTables() as $table)
 	$statement = $dbh->prepare('SELECT id FROM '.$table.' ORDER BY id DESC LIMIT 1;');
 	$statement->execute();
 	$item = $statement->fetchObject();
+	if ($item->id < $max)
+	{
+		echo "$table: id {$item->id} is less than $max (skip)\n";
+		continue;
+	}
 	echo $table.': Delete older IDs than '.($item->id - $max)."\n";
 
 	// Delete in chunks up to 5000*1000 in total, important not to leave gaps
 	for ($i = 0; $i < 5000; $i++) {
 		$newest = $item->id - $max - ($i * 1000);
-		$oldest = $newest - 1000;
+		if ($newest < 0)
+			break;
+		$oldest = max($newest - 1000, 0);
 		echo "$table: Delete between $newest and $oldest\n";
+
 		$statement = $dbh->prepare('DELETE FROM '.$table.' WHERE id < :newest AND id >= :oldest;');
 		$statement->execute(array(':newest' => $newest, ':oldest' => $oldest));
 		$deleted = $statement->rowCount();
